requirements.txt
streamlit==1.32.0
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
scikit-learn==1.3.2 